Improve Cart component and style adjustments
Refactoring has been done on the Cart component to display proper meal names instead of just their IDs. In addition, a ref has been added to the CartModal component for future use. Lastly, a newline was added to the end of index.css and no functionality or style changes were done, just a good practice for linting and readability reasons.

Co-authored-by: AgustinaScapusio <agustina.scapusio.99@gmail.com>