Update modal handling and styling in cart component
The commit replaces the showModal function with toggleVisibility for better modal handling in NavBar.tsx. The changes in CartModal.tsx and Cart.tsx reflect adjustments to the modal's styling, making it both more functional and visually appealing. Adjustments in index.html ensure better placement of the modal on the page.

Co-authored-by: AgustinaScapusio <agustina.scapusio.99@gmail.com>