Implement signOut function and refactor login
The commit replaces the fetchToken function with a signOut function that utilizes the removeToken function from a newly created localstorage utility file. This utility file includes functions that manage token in the local storage. The loginSlice state is then updated accordingly to reflect these changes, eliminating the loading status for fetching tokens and adding the same for signOut. Also, minor refactoring in the postLogin function has been done.

Co-authored-by: AgustinaScapusio <agustina.scapusio.99@gmail.com>